rename answersapi folder to searchapi
