+++ src/models/searchapi/ApiResponse.ts
