+++ src/models/searchapi/SearchError.ts
