+++ src/validation/ApiResponseValidator.ts
@@ -1,5 +1,5 @@
-import { ApiResponse } from '../models/answersapi/ApiResponse';
-import { SearchError } from '../models/answersapi/SearchError';
+import { ApiResponse } from '../models/searchapi/ApiResponse';
+import { SearchError } from '../models/searchapi/SearchError';
 
 /**
  * Determines whether or not an API response can be used to construct a search-core response.
